fix: set required flags
This is controversial. Traditional design is that flags are only used
for optional things, and positional arguments are used for required
things. However, as it stands now, flags are used for everything and
unhandled errors happen if those required values are missing.

I may revisit this before 1.0.0 to adopt positional arguments for
required values. For now, this will do.

Signed-off-by: Mark Cornick <mcornick@mcornick.com>